requirements-updated.txt
kafka-python==2.0.2
boto3==1.28.85
snowflake-connector-python==3.5.0
dbt-core==1.7.4
dbt-snowflake==1.7.1
requests==2.31.0
python-dotenv==1.0.0
apache-airflow==2.9.3
